# [WIP] Pre-allocate slices to prevent resizing at runtime