It was noticed that some services were spending a lot of time calling `runtime.growSlice` on the `wire.Field` slice. Additionally, most of the time on these slice resize calls were to `runtime.mallocgc`, and not to `runtime.memmove`. This hints the extra calls are commonly due to resizing smaller slice lengths.

This patch pre-allocates the slice to optimize the common case (until PR #469 is merged). Did the same in the case of reading large fields in `readBytes`.

Benchmark results:
```
name                                           old time/op    new time/op    delta
_readStruct/small_struct_with_64_fields-12     8.19µs ± 2%    6.86µs ± 4%    -16.16%  (p=0.008 n=5+5)
_readStruct/large_struct_with_512_fields-12    65.6µs ± 3%    57.2µs ± 1%    -12.74%  (p=0.008 n=5+5)

name                                           old alloc/op   new alloc/op   delta
_readStruct/small_struct_with_64_fields-12     12.0kB ± 0%    12.3kB ± 0%     +2.81%  (p=0.008 n=5+5)
_readStruct/large_struct_with_512_fields-12    98.0kB ± 0%    86.0kB ± 0%    -12.20%  (p=0.008 n=5+5)

name                                           old allocs/op  new allocs/op  delta
_readStruct/small_struct_with_64_fields-12       7.00 ± 0%      1.00 ± 0%    -85.71%  (p=0.008 n=5+5)
_readStruct/large_struct_with_512_fields-12      10.0 ± 0%       3.0 ± 0%    -70.00%  (p=0.008 n=5+5)
```